@thedist 2025-02-06
One of the reasons I didn't want it to remain an array is the performance improvements I'm testing in companion module base involves using a Map rather than an object (but trying to send a Map over IPC results in an empty object `{}`, and transforming it to a Map rather than an Object on Companions side didn't show much in the way of performance gains (at least not with the current setup).