@jswalden 2025-02-07
Letting the internal value encoding (or encoding constrained by IPC functionality) bleed into the user-facing UI seems like kind of a mistake.  If you're going to require modules to affirmatively opt in to the newer, faster UI, maybe it makes sense to add a new, parallel UI -- and then you can implement the old slow UI in terms of the new one.  (And by "new UI" maybe that also includes separating the variable-setting case from the variable-deleting case int two new functions.)