@thedist 2025-02-07
Converted to Draft while I'm working on some changes on the companion-module-base to assist in this.