@jswalden 2025-02-07
> What user-facing UI are you talking about?

Sorry, by "user" I meant "module developer" as opposed to "Companion implementer".  Didn't think of the obvious confusion that word would have here!  In my head modules use Companion, so they are "users", even if that's plainly not the best name to be thinking of here.