@thedist 2025-02-07
Well in terms of integration for Module Developers, I personally don't see what benefit it would have to split this into multiple ways for developers to have their module send variables to Companion.

Currently, module developers use `instance.setVariableValues` and pass it an object.
From there companion-module-base (still within the instance of that module) iterates over it and makes an array of objects which is then sent to Companion over IPC.
Companion takes that array and turns back into an object, and processes the variables internally within Companion.

The changes I'm working on are to 1. lower the overhead between companion-module-base and Companion, this would be entirely transparent to Module Developers, you'd still set your variables exactly the same way. and 2. updating `setVariableValues` within companion-module-base to support a Map, as on the module developers side this can have performance gains when generating the variables and values before they are even sent to `setVariableValues`, but currently a Map has to first be changed to an Object before passing to `setVariableValues` which has some overhead itself.

What will this mean for Module Developers?
if you want to send an object to `setVariableValues` like all modules currently do, that'll work without changes.
If you want to send an object to `setVariableValues`, but have a version of Companion and use companion-module-base with these changes, you'll have performance benefits without having to do anything.
If you want to send a Map to `setVariableValues`, that'll work too with a version of companion-module-base that supports it.
If you want to delete variables, again, use `setVariableValues` the same way it has been already and just use `undefined` as a value, how they are deleted would all be handled behind the scenes as Module Developers don't need to get into the specifics of it.

So my plans are to be full backwards compatible, no need for separate functions to send variables using the new way, no need for a separate function to delete variables.